third_party/blink/web_tests/virtual/css-gap-decorations/external/wpt/css/css-multicol/parsing/column-rule-color-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['column-rule-color'] = "green blue" should not set the property value
  assert_equals: expected "" but got "green blue"
Harness: the test ran to completion.
